figure out another GL_FLUSH bit
It appears that Vivante hw has two texture caches that need to be
flushed separately, one for fragment shaders one for vertex shaders.

This solves a problem with running cubemap_sphere after displacement;
it seemed that some leftover cache from using a texture in displacement
caused the texture in cubemap_sphere (which is only 1x1x6) to be messed
up.