driver: allocate TS only when needed
TS (Tile Status) keeps track of the clear status per tile.
Thus it is only needed when a resource is bound to a surface,
not for normal textures.
This used to be switched on PIPE_BIND_RENDER_TARGET, but mesa
creates *every* texture with this bit set if possible so we were
wasting some precious consecutive memory here.