driver: improve varyings and flat shading handling
According to the gallium documentation when flat shading
is enabled, fragment shader inputs with
TGSI_SEMANTIC_COLOR should not be interpolated. However, all other
varyings must be.

This was not implemented correctly in our varyings conversion pass which
caused problems with the blitter. This behavior has been fixed
in this commit.

In the process I figured out the meaning of bit 0 of pa_attributes.
The meanings of the other bits are still unknown.