Avoid leaking context buffers in etna_free()
Testing on Dove reveals that the context buffers are leaked each time
Etnaviv is closed down.  This is caused by the ifdef which surrounds
the call to enta_bo_del() always being false, and so the code to free
this buffer is never compiled.

Fix the ifdef, and also fix the code so that it builds.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>