Direct debugging output via stderr not stdout
Avoid writing to stdout; depending on the program using etnaviv, stdout
may not be the console, and, in the case of programs such as Xorg, may
not be something which should receive any output.

As the C library uses stderr internally for assert(), it is reasonable
to expect that this will be suitable for debug output.  Therefore,
direct etnaviv debugging output to stderr.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>