Properly handle ioctl() returning EINTR
Jon Nettleton's kernel galcore driver can return EINTR while waiting
for user signals.  With this galcore version, if we are waiting on a
command buffer galcore signal (to indicate that the buffer is now
available for re-use), and we are sent a user signal (eg, SIGALRM),
then ioctl() may return EINTR.

This causes etnaviv to believe that the galcore signal was received,
and it is safe to proceed with the next command buffer.  In reality,
this can cause the GPU to scribble over memory which it should not,
and can cause the entire platform to lock up.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>