Add support to permit mapping memory with protection info
The standard galcore user memory mapping function fails when asked to
map read-only memory, such as X shmem buffers.  The reason these fail
is because get_user_pages() is always attempted with "write" mode
enabled, which is not permissible with read-only mappings.

To allow this to work (so that Xv can be implemented efficiently) we
must be able to map read-only buffers.  This adds an extension to
etnaviv, found in my Dove galcore driver and Jon Nettleton's iMX6
kernels which permit this without changing the existing galcore APIs.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>